Bug fixes and MCP23017 Library
Remade the library for the MCP23017 using its binary register values. Functions as planned.

Fixed I2C ADC library to read float values, and added support for pH reading with the new ADC.

Turbidity was updated to have i2c adc and the port will be assigned to 1.

+++ include/I2Cadc.h
@@ -8,7 +8,7 @@ class I2Cadc   {
     public:
         static I2Cadc& Get();
         void begin();
-        int16_t readADC(int pin);
+        float readADC(int pin);
         void readDifferential();
         void setGain(adsGain_t gain);
